+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_sqlite_enum_as_string.dart
@@ -9,10 +9,9 @@ Future<SqliteEnumAsString> _$SqliteEnumAsStringFromSqlite(
     {required SqliteProvider provider,
     SqliteFirstRepository? repository}) async {
   return SqliteEnumAsString(
-      someField: MyEnum.values.firstWhere(
-          (v) => v.toString().split('.').last == data['some_field'] as String),
-      someFields: jsonDecode(data['some_fields'] ?? []).map((d) => MyEnum.values
-          .firstWhere((v) => v.toString().split('.').last == d as String)))
+      someField: MyEnum.values.byName(data['some_field'] as String),
+      someFields: jsonDecode(data['some_fields'] ?? []).whereType<String>()
+          .map(MyEnum.values.byName)
     ..primaryKey = data['_brick_id'] as int;
 }